+++ src/remix/sdk-setup.ts
@@ -334,7 +334,7 @@ export async function updateBuildScript(args: {
     : 'remix build';
 
   const instrumentedBuildCommand =
-    `${buildCommand} --sourcemap && npx --package=@sentry/remix sentry-upload-sourcemaps --org ${args.org} --project ${args.project}` +
+    `${buildCommand} --sourcemap && npx @sentry/remix --upload-sourcemaps --org ${args.org} --project ${args.project}` +
     (args.url ? ` --url ${args.url}` : '') +
     (args.isHydrogen ? ' --buildPath ./dist' : '');