+++ src/sourcemaps/tools/remix.ts
@@ -50,13 +50,13 @@ In case you already tried the wizard, we can also show you how to configure your
       `Build your app with ${chalk.cyan(
         'remix build --sourcemap',
       )}, then upload your source maps using ${chalk.cyan(
-        'npx --package=@sentry/remix sentry-upload-sourcemaps',
+        'npx @sentry/remix --upload-sourcemaps',
       )} cli tool.`,
     );
 
     clack.log.step(
       `You can add ${chalk.cyan(
-        'npx --package=@sentry/remix sentry-upload-sourcemaps',
+        'npx @sentry/remix --upload-sourcemaps',
       )} to your build script in ${chalk.cyan('package.json')} like this:`,
     );
 
@@ -67,7 +67,7 @@ In case you already tried the wizard, we can also show you how to configure your
     clack.log.step(`or run it manually after building your app.
 
 To see all available options, run ${chalk.cyan(
-      'npx --package=@sentry/remix sentry-upload-sourcemaps --help',
+      'npx @sentry/remix --upload-sourcemaps --help',
     )}
 `);
 
@@ -84,7 +84,7 @@ To see all available options, run ${chalk.cyan(
 const codeSnippet = chalk.gray(`
 "scripts": {
   ${chalk.greenBright(
-    '"build": "remix build --sourcemap && npx --package=@sentry/remix sentry-upload-sourcemaps"',
+    '"build": "remix build --sourcemap && npx @sentry/remix --upload-sourcemaps"',
   )};
 }
 `);